+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed an issue where Git based template locations would fail if the initial `git clone` failed.
+
 ## 1.32.6
 
 #### 🐞 Fixes